ts_tunnel: factor Queue out to its own module
Updates #339
